pyproject.toml
[project]
name = "marimba"
version = "1.1.0"
description = "A Python framework for structuring, processing, packaging and distributing FAIR scientific image datasets"
authors = [
    { name = "Chris Jackett", email = "chris.jackett@csiro.au" },
    { name = "Kevin Barnard", email = "kbarnard@mbari.org" },
]
requires-python = ">=3.12,<3.15"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "typer>=0.16.0,<0.17",
    "rich>=14.1.0,<15",
    "PyYAML~=6.0",
    "ifdo>=1.3.0,<2",
    "pillow>=11.3.0,<12",
    "opencv-python-headless>=4.11.0,<5",
    "gitpython>=3.1.45,<4",
    "boto3>=1.40.6,<2",
    "staticmap>=0.5.7,<0.6",
    "PyExifTool>=0.5.6,<1",
    "tabulate>=0.9.0,<0.10",
    "av>=15.0.0,<16",
    "distlib>=0.4.0,<0.5",
    "numpy>=2.2.5,<2.3.0",
    "botocore>=1.40.6",
    "requests>=2.32.4",
    "psutil>=7.0.0,<8",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
marimba = "marimba.main:marimba_cli"

[project.urls]
"Homepage" = "https://github.com/csiro-fair/marimba"
"Documentation" = "https://github.com/csiro-fair/marimba/tree/main/docs"
"Bug Tracker" = "https://github.com/csiro-fair/marimba/issues"

[dependency-groups]
dev = [
    # Core Project Management
    "hatch>=1.14.1,<2",

    # Formatting & Linting
    "black>=25.1.0,<26",
    "ruff>=0.12.8,<0.13",

    # Static Type Checking
    "mypy>=1.17.1,<2",
    "types-pyyaml>=6.0.12.20240311,<7",
    "types-tabulate>=0.9.0.20240106,<0.10",
    "types-requests>=2.32.0.20240622,<3",

    # Testing
    "pytest>=8.0.0,<9",
    "pytest-cov>=6.2.1,<7",
    "pytest-mock>=3.15.0,<4",

    # Code & Dependency Analysis
    "bandit>=1.7.8,<2",
    "deptry>=0.23.0",
    "vulture>=2.14.0,<3",

    # Git Hooks & Automation
    "pre-commit>=4.3.0,<5",

    # Packaging & Publishing
    "twine>=6.1.0,<7",
]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.env
  | \.venv
  | build
  | dist
  | temp
)/
'''

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["marimba"]

[tool.hatch.build.targets.wheel.force-include]
"marimba/py.typed" = "marimba/py.typed"

[tool.deptry.package_module_name_map]
pillow = ["PIL"]
PyYAML = ["yaml"]
gitpython = ["git"]
opencv-python-headless = ["cv2"]
PyExifTool = ["exiftool"]
